Implement default autoregister setting
This setting decides whether newly created communities will be added to the
autoregister target list.  When the module is installed, this setting defaults
to false and no communities are initially added to the autoregister target
list.  If the setting is true, when a new community is created, it will be
added to the autoregister target list and all users will be added to its
member group.