Fix upgrade to 1.0.1.
The previous upgrade should have gone from 1.0.0 to 1.0.1, but instead
went to 1.1.0.  To property semantic version, this should be an upgrade
to 1.0.1.  Additionally, there is a new requirement in Midas that each
version of a module provide a full database install script for the
plugin at that version state, so this has been added.  This should fix
the problem where the module could have been upgraded before but not
installed and upgraded in one go.  Now the module will install correctly
at version 1.0.1